- image border rendering has been corrected; - scale calculation updated;
git-svn-id: svn://svn.code.sf.net/p/simpleviewer/code@111 3ae4a26e-3d8d-4515-af9a-f071fbd64532

+++ src/pixelinfo.h
@@ -19,7 +19,7 @@ typedef struct PIXELINFO
     unsigned char* bitmap;
     int w, h, pitch;
     int bpp, format;
-    CRect<int> rc;
+    CRect<float> rc;
     float scale;
 } PixelInfo;
